t-SNE script and image on sentiment categories
Each point in tsne.png is an email. The original dataset projected was composed by the scores per sentiment category (8).
Color is given by the intensity of the color, but, if you notice the colorbar, it's quite a skewed distribution.
There in fact lots of outliers that are due to big emails like this one 'dasovich-j/sent/11919.' .
I will try to find a good criterion to remove them.